+++ src/main/kotlin/com/kennycason/nn/data/MatrixImageDecoder.kt
@@ -1,8 +1,8 @@
 package com.kennycason.nn.data
 
-import org.jblas.DoubleMatrix
+import org.jblas.FloatMatrix
 
 
 abstract class MatrixImageDecoder(val rows: Int) {
-    abstract fun decode(data: DoubleMatrix): Image
+    abstract fun decode(data: FloatMatrix): Image
 }
