+++ src/main/kotlin/com/kennycason/nn/extensions.kt
@@ -1,17 +1,17 @@
 package com.kennycason.nn
 
-import org.jblas.DoubleMatrix
+import org.jblas.FloatMatrix
 
 inline fun <reified INNER> array2d(sizeOuter: Int, sizeInner: Int,
                                    noinline innerInit: (Int)->INNER): Array<Array<INNER>> = Array(sizeOuter) { Array<INNER>(sizeInner, innerInit) }
 
-fun DoubleMatrix.apply(fn : (x: Double) -> Double) = applyi(fn, this, DoubleMatrix(rows, columns))
+fun FloatMatrix.apply(fn : (x: Float) -> Float) = applyi(fn, this, FloatMatrix(rows, columns))
 
-fun DoubleMatrix.applyi(fn : (x: Double) -> Double) : DoubleMatrix {
-    return applyi(fn, this, DoubleMatrix(rows, columns))
+fun FloatMatrix.applyi(fn : (x: Float) -> Float) : FloatMatrix {
+    return applyi(fn, this, FloatMatrix(rows, columns))
 }
 
-fun DoubleMatrix.applyi(fn : (x: Double) -> Double, other: DoubleMatrix, result: DoubleMatrix) : DoubleMatrix {
+fun FloatMatrix.applyi(fn : (x: Float) -> Float, other: FloatMatrix, result: FloatMatrix) : FloatMatrix {
     other.data.forEachIndexed { i, d -> result.data[i] = fn(d) }
     return result
 }
