+++ src/main/kotlin/com/kennycason/nn/learning_rate/ConstantLearningRate.kt
@@ -1,6 +1,6 @@
 package com.kennycason.nn.learning_rate
 
 
-class ConstantLearningRate(private val learningRate: Double) : LearningRate {
+class ConstantLearningRate(private val learningRate: Float) : LearningRate {
     override fun apply() = learningRate
 }
